Refactor the volume ls command flagging process
