+++ pkg/api/types/volume_types.go
@@ -1,0 +1,38 @@
+/*
+   Copyright The containerd Authors.
+
+   Licensed under the Apache License, Version 2.0 (the "License");
+   you may not use this file except in compliance with the License.
+   You may obtain a copy of the License at
+
+       http://www.apache.org/licenses/LICENSE-2.0
+
+   Unless required by applicable law or agreed to in writing, software
+   distributed under the License is distributed on an "AS IS" BASIS,
+   WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+   See the License for the specific language governing permissions and
+   limitations under the License.
+*/
+
+package types
+
+import "io"
+
+type VolumeLsCommandOptions struct {
+	// Writer is the output writer
+	Writer io.Writer
+	// Only display volume names
+	Quiet bool
+	// Format the output using the given go template
+	Format string
+	// Display the disk usage of volumes. Can be slow with volumes having loads of directories.
+	Size bool
+	// Filter matches volumes based on given conditions
+	Filters []string
+	// containerd namespace
+	Namespace string
+	// Root directory of persistent nerdctl state
+	DataRoot string
+	// containerd address
+	Address string
+}
